Add files via upload
Analysis on Plateau Mediclaim Insurance. To understand customer behaviors and revenue generation within the period under review.
Which forms the main objectives of the project.
The report is in Three faces ; Html file showing the source code algorithm inbuilding the project, a pdf and PowerPoint report format for presentation to the board.